host: Define a custom boot.cc main
This will be used in the future for early display initialization on
host. Currently pw_display_driver_imgui is calling glfwInit() and
this happens in a separate thread instead of main(). On MacOS this
results in the GLFW window never starting. More info:
https://www.glfw.org/docs/3.3/intro_guide.html#thread_safety

This CL doesn't address the glfw init failure, only forking boot.cc
for pw_system on host.

Bug: b/306032950
Change-Id: I0f13d96f29b01e7a1c35c32a65d8ee768b8b686f
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/kudzu/+/176830
Reviewed-by: Armando Montanez <amontanez@google.com>