[roll third_party/pigweed] pw_presubmit: Handle missing upstream better
If we try to find the commit at the remote tracking branch but are
unsuccessful, use 'HEAD~10'. This is excessive for individual CLs, but
not overly so, and it should be enough for most stacks of CLs.

Original-Bug: b/282808936
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/177038

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 20847156966b050..dd93fa16b9de0a3
Roller-URL: https://ci.chromium.org/b/8765781238761860641
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: If8eaee90ea77741373585155a391a98f3c5ed542
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/kudzu/+/178433
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>